FastAPI/app/requirements.txt
# Web 框架
fastapi==0.109.1
uvicorn[standard]==0.23.2

# 文档处理
PyPDF2==3.1.0
docx2txt==0.8
antiword==0.0.5  # 仅在 Linux/WSL 可用，需要系统安装 antiword 命令

# PDF 转图片
pdf2image==1.16.3
pillow==10.1.0

# 异步文件操作
aiofiles==23.2.1

# URL 编码
urllib3==2.0.5

# Ollama VLM 客户端
ollama==0.0.8  # 根据你的安装版本调整

# 可选工具（开发/调试）
python-multipart==0.0.6  # FastAPI 上传文件支持
